pyproject.toml
[project]
name = "hmpps-health-ping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hmpps-sre-python-lib",
    "redis==7.1.0",
]

[tool.uv.sources]
hmpps-sre-python-lib = { url = "https://github.com/ministryofjustice/hmpps-sre-python-lib/releases/download/v1.0.0/hmpps_sre_python_lib-1.0.0-py3-none-any.whl" }
